Support dummy classifier and linear kernel SVC
Currently, catwalk does not support sklearn's DummyClassifier because it does not have feature importances. This PR changes that by writing a row with the feature 'Algorithm does not support a standard way to calculate feature importance.' to the feature_importances table whenever we were not able to find feature importances.
For SVCs with linear kernel it also uses the coef_ attribute to get feature importances.