Apply replace flag to Model metadata [Resolves #23]
At present, the replace flag has no effect on model metadata, only the model pickle file.
This is problematic, as something may have changed even if
the model hashes to the same model id, such as the source data. The replace flag is meant to
signal this, so any new data should be used.

The new behavior for 'replace' is this: the non-unique attributes, like timestamps, are updated on the model row.
Furthermore, feature importances are deleted and the new ones inserted in place.

There is some refactoring here that takes advantage of a new utility function available.
save_db_objects uses a COPY command to save SQLAlchemy database objects quickly and without requiring a Session.
Since many feature importance records might be written at one time when saving a model, this is used. A Session is still
used for adding/updating the model object itself, as using a COPY command for one row would have minimal utility and
the Session provides much more value (model_id autogeneration, and easy updating of the row without constructing a large SQL query)